camera added to the model
