+++ vis/data/cameraData.js
@@ -8,20 +8,21 @@
         },
     */
 
-let animals =[
+export let animals =[
     {"name": "Lion",
      "cameraX": 0,
      "cameraY": 5,
      "cameraZ": 0
     },
     {"name": "Camera 2",
-     "cameraX": -90,
-     "cameraY": 0,
+     "cameraX": -40,
+     "cameraY": 5,
      "cameraZ": 0,
     },
     {"name": "Camera 3",
      "cameraX": 80,
-     "cameraY": 0,
+     "cameraY": 5,
      "cameraZ": 0,
     },
 ]
+
