+++ vis/savannah/index.html
@@ -2,6 +2,7 @@
 <html lang="en">
 <head>
     <title>Savannah - NZEG</title>
+    <link rel="stylesheet" href="../styles/nav.css">
     <script src="https://aframe.io/releases/1.5.0/aframe.min.js"></script>
     <script src="https://cdn.jsdelivr.net/gh/c-frame/aframe-extras@7.2.0/dist/aframe-extras.min.js"></script>
 </head>
@@ -17,7 +18,16 @@
         //     return (`0 ${theta} 0`);
         // }
     </script>
-    <a-scene>
+    <!-- arrows for navigation -->
+    <div>
+        <button class="navigationButton navigationText" id="leftNavigationButton"><p class="center">&lt</p></button>
+        <div id="navigationIndexDiv">
+          <!--inital animal for navigation-->
+          <p id="navigationIndex">Lion</p>
+        </div>
+        <button class="navigationButton navigationText" id="rightNavigationButton"><p class="center">&gt</p></button>
+    </div>
+    <a-scene id="aframe">
         <a-assets>
         </a-assets>
             <!-- x -13.25 z -->
@@ -87,7 +97,7 @@
                 "
             >
             </a-entity>
-            <a-entity camera="" position="-5.32348 2.77253 33.85177" wasd-controls="" rotation="-4.010704565915761 -16.730367617819976 0" 
+            <a-entity id="camera" camera="" position="-5.32348 2.77253 33.85177" wasd-controls="" rotation="-4.010704565915761 -16.730367617819976 0" 
             look-controls="" aframe-injected="" data-aframe-inspector-original-camera=""></a-entity>
             
     </a-scene>
@@ -130,5 +140,6 @@
 
         el.addEventListener('animationbegin', animationCorresponder)
     </script>
+    <script type="module" src="../scripts/cameraNavigation.js"></script>
 </body>
 </html>
